Improve jsdoc summaries.
The summary is the first sentence of the method/class description.  In
some cases we're missing a period so our "summary" runs on and on.

Many of these were fixed up automatically by the
`jsdoc/require-description-complete-sentence` eslint rule, but
it has a few bugs yet (especially
https://github.com/gajus/eslint-plugin-jsdoc/issues/62 ) so we
can't yet turn it on by default.

Change-Id: I5163ed2c52b21090ce018573bb3ff06d50620d73