Add mocha test which compiles the foma language converter specifications
This compiles the foma specifications and verifies that any EXPECT
assertions are satisfied (in the spirit of [[:en:Expect]].  These
confirm that the bracketing FSTs are correct (yield reversible
transforms) and that the conversion FSTs are complete (given any valid
UTF-8 input, they never fail and they never yield ambiguous results).
In addition, some basic sanity checks on example texts are done to
prevent regressions.

Since compiling the foma specifications can take 10 minutes or so (!),
and foma is not installed on jenkins, this mocha test is not run unless the
environment variable PARSOID_TEST_FOMA is set.

Change-Id: Ia3f7f1ef2c04e0d2ca713c25b42efed4066e1eaf