Port FST.js to PHP
Does what it says on the tin. A new src/ directory is created for the
PHP classes, following what appears to be the custom in the main Parsoid
repo.

The new FST class includes a static zeroFillRightShift method that
replicates the >>> operator in JS. Unit tests are provided.

The existing language tests are also ported from JS, along with a partial
port of ReplacementMachine.js, to help ensure its correctness.

All new PHPUnit tests can be run with `composer test`.

Bug: T225024
Change-Id: I37cba54813e1ec4087dd9e578e7abd02a550dbbc