Finish port of ReplacementMachine
Finishes the port of the ReplacementMachine class. Unfortunately this
means that it can no longer be tested solely within the context of this
library, since conversion depends on DOM traversal and manipulation
utilities in the main Parsoid repo. It will be imported and tested
there.

The Language tests will be moved into the main Parsoid repo, since they
will have additional Parsoid dependencies.

A new ZhReplacementMachine class used in Parsoid\Language\LanguageZh is
introduced here since the parent class lives here (and because the
porting commit there is quite large).

Change-Id: Ie61a08c200d07f6977fafa7b6bbe802b060fa713