Execute FST in depth-first order rather than breadth-first
As soon as we see an epsilon edge follow it, and store a pointer to resume
on the stack.  Previously, we were pushing edges for all epsilon edges we
saw onto the stack and defering them, bloating the stack with untaken
possibilities.

This improves performance on a 71k benchmark converting the zh-tw
localization of [[:zh:贝拉克·奥巴马]] to zh-cn from 0.64s to 0.63s.
(Presumably also improves memory footprint, though this wasn't
measured in the benchmark.)

Change-Id: Ic25c749324dfd278638e377468e52bd1c3e48106