Construct string-replacement FSTs using flag diacritics
The idea is to reduce FST complexity at the expense of runtime: flag
diacritics are executed as part of the FST but allow you to store a
limited amount of state, and thus avoid duplicating chunks of your FST.

This patch uses flag diacritics to store a "pushback" buffer; once
a match fails we go back to the start state of the FST and re-execute
using the contents of the pushback buffer.  In theory this leads to
a compact FST proportional to the size of the string-matching table
with none of the state space blowup possible in the other schemes.

In practice, the generated FSTs are still quite large.  The maximum
lookahead length is quite large for zh-hant, and the state machines
for shifting the lookahead buffer after a partial match are
conceptually simple but involve quite a large number of transitions.
Possibly being more careful about UTF-8 state could reduce the number
of transitions generated, but the number of states is still
unexpectedly large.  Needs investigation.  Also foma/hfst don't seem
to compose FSTs with flag diacritics efficiently, which doesn't help
with the main motivation for this representation, which is to allow us
to build the composed bracket-matching machines without huge demands
on CPU and memory.

Change-Id: Ibad406785d7b8b0b6047b3ca9513aee12dc3e534